Slight change to run example scripts

+++ scripts/linux/run_dev.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ $# -eq 0 ]; then
+    echo "No example to run was provided"
+    exit 1
+fi
+
 rm -rf /tmp/lib*
-javac -cp java/target/monetdbe-java-1.0-SNAPSHOT-linux.jar example/SimpleTypes.java 
-java -cp java/target/monetdbe-java-1.0-SNAPSHOT-linux.jar:example/ SimpleTypes
+javac -cp java/target/monetdbe-java-1.0-SNAPSHOT-linux.jar example/$1.java 
+java -cp java/target/monetdbe-java-1.0-SNAPSHOT-linux.jar:example/ $1
